Use compile phase to copy the cinnamon-agent.jar
The `compile` phase is more appropriated when copying Cinnamon agent jar
and possibly it is a better default/suggestion. The reasoning here is that, based
on Maven's default lifecycle:

https://maven.apache.org/guides/introduction/introduction-to-the-lifecycle.html#lifecycle-reference

And the plugin descriptor for Maven Exec Plugin:

https://github.com/mojohaus/exec-maven-plugin/blob/94570ccbf2279a3f67c95664248221d0de291b7b/pom.xml#L235-L240

Using package is inconvenient when running the tests. If we use `package`, users
are required to run:

    mvn package -DskipTests
    mvn test

And when using `compile`, everything should work as expected when running:

   mvn test